Split query module into multiple files
