+++ packages/std/src/query/bank.rs
@@ -1,0 +1,32 @@
+use schemars::JsonSchema;
+use serde::{Deserialize, Serialize};
+
+use crate::Coin;
+
+#[non_exhaustive]
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
+#[serde(rename_all = "snake_case")]
+pub enum BankQuery {
+    /// This calls into the native bank module for one denomination
+    /// Return value is BalanceResponse
+    Balance { address: String, denom: String },
+    /// This calls into the native bank module for all denominations.
+    /// Note that this may be much more expensive than Balance and should be avoided if possible.
+    /// Return value is AllBalanceResponse.
+    AllBalances { address: String },
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
+#[serde(rename_all = "snake_case")]
+pub struct BalanceResponse {
+    /// Always returns a Coin with the requested denom.
+    /// This may be of 0 amount if no such funds.
+    pub amount: Coin,
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
+#[serde(rename_all = "snake_case")]
+pub struct AllBalanceResponse {
+    /// Returns all non-zero coins held by this account.
+    pub amount: Vec<Coin>,
+}
