+++ packages/std/src/query/mod.rs
@@ -1,0 +1,100 @@
+use schemars::JsonSchema;
+use serde::{Deserialize, Serialize};
+
+#[cfg(feature = "stargate")]
+use crate::ibc::IbcQuery;
+#[cfg(feature = "stargate")]
+use crate::Binary;
+use crate::Empty;
+
+mod bank;
+mod staking;
+mod stargate;
+mod wasm;
+
+pub use bank::{AllBalanceResponse, BalanceResponse, BankQuery};
+pub use staking::{
+    AllDelegationsResponse, BondedDenomResponse, Delegation, DelegationResponse, FullDelegation,
+    StakingQuery, Validator, ValidatorsResponse,
+};
+#[cfg(feature = "stargate")]
+pub use stargate::StargateResponse;
+pub use wasm::WasmQuery;
+
+#[non_exhaustive]
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
+#[serde(rename_all = "snake_case")]
+pub enum QueryRequest<C: CustomQuery> {
+    Bank(BankQuery),
+    Custom(C),
+    Staking(StakingQuery),
+    /// A Stargate query encoded the same way as abci_query, with path and protobuf encoded Data.
+    /// The format is defined in [ADR-21](https://github.com/cosmos/cosmos-sdk/blob/master/docs/architecture/adr-021-protobuf-query-encoding.md)
+    /// The response is also protobuf encoded. The caller is responsible for compiling the proper protobuf definitions
+    #[cfg(feature = "stargate")]
+    Stargate {
+        /// this is the fully qualified service path used for routing,
+        /// eg. custom/cosmos_sdk.x.bank.v1.Query/QueryBalance
+        path: String,
+        /// this is the expected protobuf message type (not any), binary encoded
+        data: Binary,
+    },
+    #[cfg(feature = "stargate")]
+    Ibc(IbcQuery),
+    Wasm(WasmQuery),
+}
+
+/// A trait that is required to avoid conflicts with other query types like BankQuery and WasmQuery
+/// in generic implementations.
+/// You need to implement it in your custom query type.
+///
+/// # Examples
+///
+/// ```
+/// # use cosmwasm_std::CustomQuery;
+/// # use schemars::JsonSchema;
+/// # use serde::{Deserialize, Serialize};
+/// #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
+/// #[serde(rename_all = "snake_case")]
+/// pub enum MyCustomQuery {
+///     Ping {},
+///     Capitalized { text: String },
+/// }
+///
+/// impl CustomQuery for MyCustomQuery {}
+/// ```
+pub trait CustomQuery: Serialize {}
+
+impl CustomQuery for Empty {}
+
+impl<C: CustomQuery> From<BankQuery> for QueryRequest<C> {
+    fn from(msg: BankQuery) -> Self {
+        QueryRequest::Bank(msg)
+    }
+}
+
+impl<C: CustomQuery> From<C> for QueryRequest<C> {
+    fn from(msg: C) -> Self {
+        QueryRequest::Custom(msg)
+    }
+}
+
+#[cfg(feature = "staking")]
+impl<C: CustomQuery> From<StakingQuery> for QueryRequest<C> {
+    fn from(msg: StakingQuery) -> Self {
+        QueryRequest::Staking(msg)
+    }
+}
+
+impl<C: CustomQuery> From<WasmQuery> for QueryRequest<C> {
+    fn from(msg: WasmQuery) -> Self {
+        QueryRequest::Wasm(msg)
+    }
+}
+
+#[cfg(feature = "stargate")]
+impl<C: CustomQuery> From<IbcQuery> for QueryRequest<C> {
+    fn from(msg: IbcQuery) -> Self {
+        QueryRequest::Ibc(msg)
+    }
+}
